Move libteken out of the syscons directory.
I initially committed libteken to sys/dev/syscons/teken, but now that
I'm working on a console driver myself, I noticed this was not a good
decision. Move it to sys/teken to make it easier for other drivers to
use a terminal emulator.

Also list teken.c in sys/conf/files, instead of listing it in all the
files.arch files separately.


git-svn-id: http://svn.freebsd.org/base/head@196775 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f