Don't corrupt logout entries in utmp files.
An older version of the code used a structure on the stack, instead of a
pointer to the structure. It looks like I didn't adjust the parameters
of the write(2) call, causing the first four/eight bytes of the entry to
be corrupted, instead of writing the entire entry to disk.


git-svn-id: http://svn.freebsd.org/base/head@200996 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f